backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
fastapi==0.136.1
h11==0.16.0
idna==3.14
pydantic==2.13.4
pydantic_core==2.46.4
python-dotenv==1.2.2
requests==2.34.0
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.7.0
uvicorn==0.46.0
